requirements.txt
asgiref==3.5.0
asn1crypto==1.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
crypto==1.4.1
cryptography==36.0.1
Django==4.0.2
djangorestframework==3.13.1
http-ece==1.1.0
idna==3.3
Naked==0.1.31
oscrypto==1.2.1
protobuf==3.19.4
push-receiver @ git+https://github.com/vivedp/push_receiver.git@9ad4a58b389cf8e6499eec59c061be755d29369f
pycparser==2.21
pycryptodome==3.4.3
pytz==2021.3
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
sqlparse==0.4.2
urllib3==1.26.8

aioapns~=2.1
PyJWT~=2.3.0